Move TextFieldWidget from components to CMS feature
