+++ packages/airview-cms/src/components/index.js
@@ -1,2 +1,1 @@
 export { FontAwesomeSvgIcon } from "./font-awesome-svg-icon";
-export { TextFieldWidget } from "./text-field-widget";
